Invalid integer types caused incorrect time on 32-bit

+++ src/get_time.cpp
@@ -43,7 +43,8 @@ uint64_t get_time_since_epoch_ms() {
 uint64_t get_time_since_epoch_ms() {
   struct timeval tv;
   gettimeofday(&tv, NULL);
-  return tv.tv_sec * 1000 + tv.tv_usec / 1000;
+  return static_cast<uint64_t>(tv.tv_sec)  * 1000 +
+         static_cast<uint64_t>(tv.tv_usec) / 1000;
 }
 
 #else
@@ -51,7 +52,8 @@ uint64_t get_time_since_epoch_ms() {
 uint64_t get_time_since_epoch_ms() {
   struct timespec ts;
   clock_gettime(CLOCK_REALTIME, &ts);
-  return ts.tv_sec * 1000 + ts.tv_nsec / 1000000;
+  return static_cast<uint64_t>(ts.tv_sec)  * 1000 +
+         static_cast<uint64_t>(ts.tv_nsec) / 1000000;
 }
 
 #endif
